Finished GoalAuto and MidAuto. Changed Talon to WPI_TalonSRX

+++ infinite-recharge-robot/src/main/java/frc/robot/commands/CommandFactory.java
@@ -19,7 +19,7 @@ public class CommandFactory {
             driveSystem
           );
     }
-    public static Command angleTurnCommand(double speed, int angle, Direction direction, DriveSystem driveSystem) {
+    public static Command angleTurnCommand(double speed, double angle, Direction direction, DriveSystem driveSystem) {
         return new FunctionalCommand(
             ()-> driveSystem.resetAngle(), 
             ()-> driveSystem.turn(speed, direction),
